+++ anymaps-base/src/main/java/net/doo/maps/model/MarkerOptions.java
@@ -18,7 +18,7 @@ public class MarkerOptions {
 	private float anchorU = 0f;
 	private float anchorV = 0f;
 	private BitmapDescriptor icon;
-	private float z = 0f;
+	private int z = 0;
 
 	/**
 	 * @param alpha alpha-level of the marker. In range [0..1]. Default value is 1.
@@ -71,7 +71,7 @@ public class MarkerOptions {
 		return this;
 	}
 
-	public MarkerOptions z(float z) {
+	public MarkerOptions z(int z) {
 		this.z = z;
 		return this;
 	}
@@ -121,7 +121,7 @@ public class MarkerOptions {
 	/**
 	 * @see #z(float)
 	 */
-	public float getZ() {
+	public int getZ() {
 		return z;
 	}
